% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detectTop10.R
\name{detectTop10}
\alias{detectTop10}
\title{detectTop10}
\usage{
detectTop10(P, QTN.position)
}
\arguments{
\item{P}{vector of p-values of all genes}

\item{QTN.position}{position of the simulated QTNs}
}
\value{
number of QTN detected in top 10
}
\description{
Finds how many simulated genes made it into top ten
}
