% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/G2P.R
\name{G2P}
\alias{G2P}
\title{G2P}
\usage{
G2P(X, h2, alpha, NQTN, distribution)
}
\arguments{
\item{X}{A Genotype Matrix (n x m)}

\item{h2}{Heritability}

\item{alpha}{affect of the qtns}

\item{NQTN}{Number of QTNS to be simulated}

\item{distrubution}{which distribution to use, only "norm" is available}
}
\value{
simulated phenotype list
}
\description{
Genotype 2 Phenotype function.
Taken from Lecture, using for simulation.
}
