% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2dpcplots.R
\name{graphPCA}
\alias{graphPCA}
\title{graphPCA}
\usage{
graphPCA(allPC)
}
\value{

}
\description{
Plots the Scree/Elbow plot,the Proprtion of the variance explained by the PC's and the Cumulative Proportion of the variance explained by the PC's
}
