% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GWASbyCor.R
\name{GWASbyCor}
\alias{GWASbyCor}
\title{GWAS by Cor}
\usage{
GWASbyCor(X, Y)
}
\arguments{
\item{X}{A Genotype Matrix (n x m)}

\item{Y}{A Phenotype Matrix (n x 1)}
}
\value{
P A P-value vector (n) representing importance of each gene
}
\description{
Performs GWAS by Correlation
}
