% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcCoVariance.R
\name{pcCoVariance}
\alias{pcCoVariance}
\title{pcCoVariance}
\usage{
pcCoVariance(PC, C, threshold)
}
\arguments{
\item{PC}{(n x p) p Principle components for n samples}

\item{C}{(n x c) c Covariates for n samples}
}
\value{
PC_C (n x (p + c - linear dependent)) Principle Components and Covariates which are independent
}
\description{
Determines principle components (PC) that are in linear dependent to the covariates (C).
Removes PC that are linear dependent
}
