% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcThreshold.R
\name{pcThreshold}
\alias{pcThreshold}
\title{pcThreshold}
\usage{
pcThreshold(PC, threshold = 0.014)
}
\arguments{
\item{PC}{principle components generated by prcomp}

\item{threshold}{the amount a principle component must explain to be kept in the model}
}
\value{
PC_T
}
\description{
Determines principle components to use in model based on set threshold
}
