% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_Manhatten.R
\name{plotManhatten}
\alias{plotManhatten}
\title{Plot Manhatten}
\usage{
plotManhatten(P)
}
\arguments{
\item{P}{A P-value vector (n) representing importance of each gene}
}
\description{
Plots manhatten plot of P values
}
