% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_QQ.R
\name{plotQQ}
\alias{plotQQ}
\title{Plot QQ}
\usage{
plotQQ(P, main)
}
\arguments{
\item{P}{A P-value vector (n) representing importance of each gene}

\item{main}{Name of plot}
}
\description{
Plots Q plot of P-value vector
}
