% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Run_GLM.R
\name{Run_GLM}
\alias{Run_GLM}
\title{Run_GLM}
\usage{
Run_GLM(myGD, y, PC)
}
\arguments{
\item{X}{A Genotype Matrix (n x m)}

\item{Y}{A Phenotype Matrix (n x 1)}

\item{C}{A Covariate Matrix (n x c)}
}
\value{
P A P-value vector (n) representing importance of each gene
}
\description{
Performs GLM on a Genotype, Phenotype and Covariate Matrix.
}
