% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcThreshold.R
\name{pcThreshold}
\alias{pcThreshold}
\title{Choosing PC's}
\usage{
pcThreshold(PC, threshold = 0.014)
}
\arguments{
\item{PC}{numeric : principal components generated by "prcomp".}

\item{threshold}{numeric: the chosen  amount or proportion of variance explained by the  number of  principle components to remain in the model.}
}
\value{
PC_T matrix: The  chosen number of principal components agreeing with the threshold.
}
\description{
To determine/choose the Principal component Threshold
pcThreshold() Determines principle components to use in model based on set threshold
}
